refactor: move common codes to kernel package
