+++ packages/nostr-fetch-kernel/src/channel.spec.ts
