+++ packages/nostr-fetch-kernel/src/channel.ts
