+++ packages/nostr-fetch-kernel/src/index.ts
@@ -1,0 +1,1 @@
+export {};
