+++ packages/nostr-fetch-kernel/src/nostr.spec.ts
