+++ packages/nostr-fetch-kernel/src/nostr.ts
