+++ packages/nostr-fetch-kernel/src/relayTypes.ts
