+++ packages/nostr-fetch-kernel/src/utils.spec.ts
