+++ packages/nostr-fetch-kernel/src/utils.ts
