+++ packages/nostr-fetch-kernel/tsconfig.json
@@ -1,0 +1,12 @@
+{
+  // https://github.com/tsconfig/bases/blob/main/bases/strictest.json
+  "extends": "@tsconfig/strictest/tsconfig.json",
+  "compilerOptions": {
+    "target": "ES2022",
+    "lib": ["ES2022", "DOM"],
+    "module": "ES2022",
+    "moduleResolution": "Node",
+    "noEmit": true,
+  },
+  "include": ["src/**/*.ts"],
+}
