+++ packages/nostr-fetch-kernel/vitest.config.ts
