+++ packages/nostr-fetch/build.js
@@ -8,9 +8,8 @@ const sharedBuildOptions = {
   entryPoints: ["src/index.ts"],
   outdir: DIST_DIR,
   bundle: true,
-  packages: "external",
   minify: true,
-  sourcemap: "external",
+  sourcemap: "linked",
 };
 
 const buildCJS = async () =>
