+++ packages/nostr-fetch/src/index.ts
@@ -1,2 +1,2 @@
+export { NostrEvent, eventKind } from "@nostr-fetch/kernel/nostr";
 export * from "./fetcher";
-export { NostrEvent, eventKind } from "./nostr";
